Rename xforce_motion to force_motion
Rename xforce_motion.py to force_motion.py across experiments (m3_wolff_lafreniere, m4_vector_wave, m5_liquid_crystal) and update all imports and in-code/documentation references. Also apply minor formatting and whitespace cleanups in engine3_observables, medium, and launcher files (single-line seed call, small comment updates, and a few reflows). This consolidates naming and keeps comments/docs consistent with the new module name.
