+++ openwave/xperiments/m3_wolff_lafreniere/_launcher.py
@@ -23,7 +23,7 @@ from openwave.i_o import flux_mesh, render, video
 import openwave.xperiments.m3_wolff_lafreniere.medium as medium
 import openwave.xperiments.m3_wolff_lafreniere.particle as particle
 import openwave.xperiments.m3_wolff_lafreniere.wave_engine as ewave
-import openwave.xperiments.m3_wolff_lafreniere.xforce_motion as force_motion
+import openwave.xperiments.m3_wolff_lafreniere.force_motion as force_motion
 import openwave.xperiments.m3_wolff_lafreniere.instrumentation as instrument
 
 # ================================================================
